chore - move tor package into packages/tor
Move the main library into packages/tor as a workspace package, keeping the
repo root as a workspace aggregator. Update examples, CI, and lint/typecheck
paths to use the new package layout.
