+++ examples/node-fetch/package.json
@@ -9,6 +9,7 @@
     "dev": "pm2 start --name tor-example-node-fetch --watch=../../ --no-daemon --interpreter=node --node-args=--experimental-transform-types index.ts"
   },
   "dependencies": {
+    "tor": "^0.1.0",
     "node-fetch": "^3.3.2"
   },
   "devDependencies": {}
