+++ packages/tor/package.json
@@ -1,0 +1,41 @@
+{
+  "name": "tor",
+  "version": "0.1.0",
+  "license": "MIT",
+  "type": "module",
+  "dependencies": {
+    "@noble/curves": "^1.9.7",
+    "@noble/ed25519": "^2.3.0",
+    "@noble/hashes": "^1.8.0",
+    "commander": "^14.0.2",
+    "onionoo": "^2.0.2"
+  },
+  "devDependencies": {
+    "@types/node": "^22.0.0",
+    "ava": "^6.4.1",
+    "pm2": "^6.0.14",
+    "typescript": "^5.9.3"
+  },
+  "exports": {
+    ".": "./src/index.ts",
+    "./node": "./src/node.ts",
+    "./build-circuit/chutney": "./src/build-circuit/chutney.ts",
+    "./build-circuit/mainnet": "./src/build-circuit/mainnet.ts",
+    "./cli": "./src/cli.ts"
+  },
+  "scripts": {
+    "dev": "pm2 start --no-daemon scripts/dev.config.cjs",
+    "test": "ava src/*.spec.ts",
+    "typecheck": "tsc -p tsconfig.json --noEmit",
+    "import:check": "node --experimental-transform-types scripts/import-check.ts",
+    "test:dev": "pm2 start --no-daemon scripts/test-dev.config.cjs"
+  },
+  "ava": {
+    "extensions": {
+      "ts": "module"
+    },
+    "nodeArguments": [
+      "--experimental-transform-types"
+    ]
+  }
+}
