+++ packages/tor/scripts/dev.config.cjs
