+++ packages/tor/scripts/import-check.ts
