+++ packages/tor/scripts/test-cell.ts
