+++ packages/tor/src/cli.ts
